move formatter scripts to formatter/ dir
