+++ src/formatter/formatAliasPositions.ts
