+++ src/formatter/formatCommaPositions.ts
